Create Fingerprint_db
This project uses Open cv, sqlite3, and matplotlib in order to create a ridge details (fingerprint and footprint) database. The code is structured to first create a table for the database named “fingerprints.db” with the table including the categories in figure 1 below. The first column is in integer format labeled “i” and will automatically update when data is entered. Moreover, I structured the database to create the name in TEXT format, dob (Date Of Birth) in DATE format, and the features (the ridge details data) in BLOB format which is from the reference code.